requirements.txt
pandas
requests
psycopg2
boto